+++ vhpidirect/cinterface/accarrayint/cSharedVar.h
@@ -1,0 +1,9 @@
+#include <stdio.h>
+#include <malloc.h>
+
+int sizeInt;
+int getIntArrSize();
+
+int* intArray;
+int* getIntArray_ptr();
+void freeIntArray();
