+++ vhpidirect/cinterface/accarrayint/run.sh
