--- vhpidirect/cinterface/accarrayvar/cSharedVar.h
@@ -1,9 +1,0 @@
-#include <stdio.h>
-#include <malloc.h>
-
-int sizeInt;
-int getIntArrSize();
-
-int* intArray;
-int* getIntArray_ptr();
-void freeIntArray();
